requirements.txt
numpy
scikit-learn
pythainlp
line-bot-sdk
fastapi
uvicorn[standard]
google-generativeai
fastapi
